requirements.txt
pycryptodome==3.15.0
cryptography==3.4.8